confirmation dialog implemented

+++ todo-app-angular-workspace/projects/todo-app/src/app/components/todo-item/todo-item.component.ts
@@ -3,6 +3,8 @@ import { Component, Input } from '@angular/core';
 import { ITodo } from '../../shared/models/todo';
 import { TodoListActions } from '../../shared/state/todo.actions';
 import { IState } from '../../shared/state/state';
+import { ConfirmModalService } from '../confirm-modal/confirm-modal.service';
+import { first } from 'rxjs';
 
 @Component({
   selector: 'app-todo-item',
@@ -10,16 +12,27 @@ import { IState } from '../../shared/state/state';
   styleUrls: ['./todo-item.component.scss']
 })
 export class TodoItemComponent {
-
   @Input() todo: ITodo = {} as ITodo;
 
-  constructor(private store: Store<IState>) {}
+  constructor(private store: Store<IState>, private modalService: ConfirmModalService) {}
 
   onComplete(todoId: number) {
-    this.store.dispatch(TodoListActions.completed({ todoId }));
+    this.modalService.confirm('Are you sure?', 'modal-sm')
+    .pipe(first())
+    .subscribe((confirmed) => {
+      if (confirmed) {
+        this.store.dispatch(TodoListActions.completed({ todoId }));
+      }      
+    });    
   }
 
   onRemove(todoId: number) {
-    this.store.dispatch(TodoListActions.removed({ todoId }));
+    this.modalService.confirm('Are you sure?', 'modal-sm')
+    .pipe(first())
+    .subscribe((confirmed) => {
+      if (confirmed) {
+        this.store.dispatch(TodoListActions.removed({ todoId }));
+      }      
+    }); 
   }
 }
